Fix premature exit of a modifier on first child
Problem: Modifiers can return the next item to iterate over, by
returning the next indice. When a modifiers removes the current
child, it is sometimes useful to check the now-current item
again, by returning `currentIndex - 1`. This went wrong if the
currentIndex was 0, exiting the modification of all children by
that modifier.