Change inner-word slash handling
Previously, all slashes between words were merged.  This change
reverses that and only allows slashes between words to merge their
surroundings when: a) The word before it is shorter than three
characters, and b) there is no following word or the following word
is shorter than three characters.

This removes incorrect treatment of `and/or` or `or/and` as a single
word, but does allow `w/`, `w/o`, `c/o`, `a/c`, or `km/h` (Dutch),
and others.